Better document the workarounds on Apple Silicon (#59)
The info box is more prominently placed towards the top of the
"Installation" page now. It is still on the "Releases" page in its
usual spot.

The workaround details are more in-depth now and contain an actual
code sample. To not clutter the page, the details are inside a
collapsable area which is collapsed by default. Since we have the
same information in two places, I factored it out into a React
component.
---------

Co-authored-by: Michael Haubenschild <haubenmi@users.noreply.github.com>