Fix deployment by downloading lang_docs with curl instead of typescript (#100)
The download of the lang_docs zip packages with typescript suddenly
started failing and we couldn't figure out why. The download itself
succeeded, while the downloaded files were empty.

To work around the issue, the downloading script now uses curl to do
the actual download. Typescript is still used to generate the curl
command line, as the information for the download is only available in
typescript.